requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==5.2.14
django-appconf==1.2.0
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-filter==25.2
django-imagekit==6.1.0
django-import-export==4.4.1
django-simple-history==3.11.0
django-storages==1.14.6
django-timezone-field==7.2.1
django-unfold==0.91.0
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
flower==2.0.1
greenlet==3.5.0
gunicorn==26.0.0
humanize==4.15.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
openpyxl==3.1.5
packaging==26.2
pilkit==3.0
pillow==12.2.0
playwright==1.59.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pyee==13.0.1
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2026.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.7.0
whitenoise==6.12.0
